Add explicit cast to prevent compiler warning.
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11062 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXtackLib/src/test/net/SipSubscribeClientTest.cpp
@@ -1905,7 +1905,7 @@ public:
          // The refresh manager should re-SUBSCRIBE
          // Wait for the next notify request and subscribe response
          int secondMessageWait = 60;
-         int resendTimeout = 0.55 * secondMessageWait;
+         int resendTimeout = (int) (0.55 * secondMessageWait);
          if(resendTimeout < 40)
          {
             resendTimeout = 40;
